requirements.txt
# API
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
python-multipart>=0.0.6

# Config
pydantic-settings>=2.1.0

# HTTP client
httpx>=0.26.0

# Extraction
pypdf>=4.0.0
python-docx>=1.1.0
openpyxl>=3.1.2
python-pptx>=0.6.23
Pillow>=10.2.0
pytesseract>=0.3.10

# Vector store & embeddings
chromadb>=0.4.22
mistralai>=1.0.0
